chore: unify onViewUpdate calls
