+++ src/scene/sprite-nine-slice/NineSliceSprite.ts
@@ -1,6 +1,6 @@
 import { Texture } from '../../rendering/renderers/shared/texture/Texture';
 import { deprecation, v8_0_0 } from '../../utils/logging/deprecation';
-import { ViewContainer } from '../view/View';
+import { ViewContainer } from '../view/ViewContainer';
 import { NineSliceGeometry } from './NineSliceGeometry';
 
 import type { Size } from '../../maths/misc/Size';
@@ -291,21 +291,10 @@ export class NineSliceSprite extends ViewContainer implements View
         return this._texture.height;
     }
 
-    protected onViewUpdate()
+    protected override onViewUpdate()
     {
-        this._didViewChangeTick++;
-
         this._didSpriteUpdate = true;
-
-        if (this.didViewUpdate) return;
-        this.didViewUpdate = true;
-
-        const renderGroup = this.renderGroup || this.parentRenderGroup;
-
-        if (renderGroup)
-        {
-            renderGroup.onChildViewUpdate(this);
-        }
+        super.onViewUpdate();
     }
 
     /**
