+++ src/scene/sprite/Sprite.ts
@@ -1,7 +1,7 @@
 import { ObservablePoint } from '../../maths/point/ObservablePoint';
 import { Texture } from '../../rendering/renderers/shared/texture/Texture';
 import { updateQuadBounds } from '../../utils/data/updateQuadBounds';
-import { ViewContainer } from '../view/View';
+import { ViewContainer } from '../view/ViewContainer';
 
 import type { Size } from '../../maths/misc/Size';
 import type { PointData } from '../../maths/point/PointData';
@@ -221,22 +221,12 @@ export class Sprite extends ViewContainer
         bounds.addFrame(_bounds.minX, _bounds.minY, _bounds.maxX, _bounds.maxY);
     }
 
-    public override onViewUpdate()
+    protected override onViewUpdate()
     {
-        this._didViewChangeTick++;
-
         this._didSpriteUpdate = true;
         this._sourceBoundsDirty = this._boundsDirty = true;
 
-        if (this.didViewUpdate) return;
-        this.didViewUpdate = true;
-
-        const renderGroup = this.renderGroup || this.parentRenderGroup;
-
-        if (renderGroup)
-        {
-            renderGroup.onChildViewUpdate(this);
-        }
+        super.onViewUpdate();
     }
 
     protected override _updateBounds()
